LocalColorExtractor: Disable local colors when using system presets
Similar to Pixel Launcher, local colors should be disabled when using a
preset color because presets will most likely differ from the wallpaper.
This is necessary for WallpaperPickerGoogle and Repainter to work properly.

Reloading works fine as LocalColorExtractor is recreated on
configuration changes.

statix edits: also update style to VIBRANT

Change-Id: I8442b60f8b2359d6d2a0e4007e96c525f3cedc96
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>